requirements.txt
websocket-client
requests
python-telegram-bot
schedule
